+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/background.tsx
