+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/purpleBear.tsx
