+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/squiggle.tsx
