+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/teamPicture.tsx
