+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/text.tsx
