+++ apps/main/src/app/lib/Assets/SVG/AboutLandingAssets/yellowBear.tsx
