+++ apps/main/src/app/lib/Assets/SVG/FAQAssets/FAQCarousel.tsx
