+++ apps/main/src/app/lib/Assets/SVG/FAQAssets/FAQFireworks.tsx
