+++ apps/main/src/app/lib/Assets/SVG/KeynoteAssets/KeynoteSpeaker.tsx
