+++ apps/main/src/app/lib/Assets/SVG/KeynoteAssets/guestphoto.tsx
