+++ apps/main/src/app/lib/Assets/SVG/OurTeamPageAssets/AboutUs.tsx
