+++ apps/main/src/app/lib/Assets/SVG/OurTeamPageAssets/hero.tsx
