+++ apps/main/src/app/lib/Assets/SVG/OurTeamPageAssets/heroBackground.tsx
