+++ apps/main/src/app/lib/Assets/SVG/ProjectsPageAssets/OrangeFirework.tsx
