+++ apps/main/src/app/lib/Assets/SVG/ProjectsPageAssets/PinkFirework.tsx
