+++ apps/main/src/app/lib/Assets/SVG/ProjectsPageAssets/ProjectBackground.tsx
