+++ apps/main/src/app/lib/Assets/SVG/ProjectsPageAssets/YellowFirework.tsx
