+++ apps/main/src/app/lib/Assets/SVG/SponsorPageAssets/SponsorBackground.tsx
