+++ apps/main/src/app/lib/Assets/SVG/SponsorPageAssets/SponsorText.tsx
