+++ apps/main/src/app/lib/Assets/SVG/SponsorPageAssets/SponsorTicket.tsx
