+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/CloudBackground.tsx
