+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/DesktopTopBox.tsx
