+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/FullCloudBackground.tsx
