+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/ImpactHeader.tsx
