+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/ImpactSectionBackground.tsx
