+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/SponsorUsTicket.tsx
