+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/Ticket.tsx
