+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/TicketBooth.tsx
