+++ apps/main/src/app/lib/Assets/SVG/SponsorUsPageAssets/TopClouds.tsx
