+++ apps/main/src/app/lib/Assets/SVG/StatsAssets/TicketBooth.tsx
