+++ apps/main/src/app/lib/Assets/SVG/StatsAssets/TicketsBack.tsx
