+++ apps/main/src/app/lib/Assets/SVG/StatsAssets/TicketsFront.tsx
