--- apps/main/src/app/lib/Assets/SVG/banner.tsx
@@ -1,35 +1,0 @@
-import React from "react";
-
-const Banner: React.FC = () => {
-  return (
-    <svg
-      width="597"
-      height="93"
-      viewBox="0 0 597 93"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M2.28232 37.9013C1.18703 35.1304 3.22915 32.1273 6.20871 32.1273H85.2844C87.6162 32.1273 89.5064 34.0176 89.5064 36.3493V86.7721C89.5064 89.1038 87.6162 90.9941 85.2844 90.9941H6.37191C3.34895 90.9941 1.30549 87.9101 2.4839 85.1263L11.3592 64.1598C11.7908 63.1401 11.8046 61.9917 11.3976 60.9619L2.28232 37.9013Z"
-        fill="#94181A"
-      />
-      <path
-        d="M84.58 90.648L54.3562 82.2767C49.955 81.0576 50.3538 74.6904 54.8728 74.0302L85.0966 69.6146C87.6443 69.2424 89.929 71.2175 89.929 73.7922V86.5792C89.929 89.37 87.2696 91.393 84.58 90.648Z"
-        fill="#68090B"
-      />
-      <path
-        d="M594.187 37.9482C595.283 35.1773 593.241 32.1742 590.261 32.1742H511.185C508.854 32.1742 506.963 34.0644 506.963 36.3962V86.819C506.963 89.1507 508.854 91.041 511.185 91.041H590.098C593.121 91.041 595.164 87.957 593.986 85.1731L585.111 64.2066C584.679 63.187 584.665 62.0385 585.072 61.0088L594.187 37.9482Z"
-        fill="#94181A"
-      />
-      <path
-        d="M512.312 90.6949L542.535 82.3236C546.937 81.1045 546.538 74.7373 542.019 74.0771L511.795 69.6614C509.247 69.2892 506.963 71.2644 506.963 73.8391V86.6261C506.963 89.4169 509.622 91.4398 512.312 90.6949Z"
-        fill="#68090B"
-      />
-      <path
-        d="M50.9629 13.8682C50.9629 11.643 52.6716 9.80167 54.8925 9.66178C79.1867 8.13148 210.753 0.106227 298.318 0.0882568C386.36 0.0701892 518.656 8.12855 543.031 9.66192C545.252 9.80164 546.963 11.6432 546.963 13.8685V77.6659C546.963 80.0744 545.013 81.9985 542.607 81.8908C517.15 80.7508 385.733 75.0623 298.318 75.0711C211.381 75.0799 80.6955 80.7511 55.3176 81.8903C52.9113 81.9984 50.9629 80.0741 50.9629 77.6654V13.8682Z"
-        fill="#CC322D"
-      />
-    </svg>
-  );
-};
-export default Banner;
