--- apps/main/src/app/lib/Assets/SVG/popcorn.tsx
@@ -1,245 +1,0 @@
-import React from "react";
-const Popcorn: React.FC = () => {
-  return (
-    <svg
-      className="w-40 h-40"
-      width="96"
-      height="130"
-      viewBox="0 0 96 130"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <mask
-        id="mask0_707_15389"
-        style={{ maskType: "alpha" }}
-        maskUnits="userSpaceOnUse"
-        x="8"
-        y="7"
-        width="69"
-        height="54"
-      >
-        <path
-          d="M8.92028 53.6819C10.2883 57.6049 16.484 60.4853 16.484 60.4853L75.9951 41.5003C75.9951 41.5003 77.3531 37.9514 76.8366 35.6564C76.302 33.2809 75.2382 31.959 73.3883 30.6009C71.8916 29.5022 69.0669 29.0804 69.0669 29.0804C69.0669 29.0804 69.8674 27.8202 70.1012 26.873C70.6832 24.5155 70.3844 22.2398 68.6097 20.6368C66.8453 19.043 62.9764 20.3804 62.9764 20.3804C62.9764 20.3804 62.7586 17.453 61.9014 15.8858C60.4375 13.2095 58.4075 11.8736 55.6902 11.9069C53.1926 11.9376 50.2766 15.3994 50.2766 15.3994C50.2766 15.3994 47.9 12.387 46.0357 10.9492C43.5008 8.99443 41.7662 7.16645 38.8214 7.56662C36.6977 7.85522 35.4359 8.66267 34.1018 10.4773C32.5348 12.6088 32.5394 17.4439 32.5394 17.4439C32.5394 17.4439 30.5698 15.6262 29.0913 15.0019C26.6933 13.9892 24.9235 13.8071 22.6564 14.8778C20.5075 15.8927 19.5393 17.3001 18.5414 19.6633C17.3512 22.4822 18.0896 27.7899 18.0896 27.7899C18.0896 27.7899 15.5775 26.8086 14.1221 27.3133C11.0001 28.396 11.5466 32.7982 12.2061 36.5028C12.6254 38.858 14.638 42.0983 14.638 42.0983C14.638 42.0983 11.067 42.82 9.66309 44.5949C7.6053 47.1963 7.73272 50.2765 8.92028 53.6819Z"
-          fill="#F2E06F"
-        />
-      </mask>
-      <g mask="url(#mask0_707_15389)">
-        <path
-          d="M8.92028 53.6819C10.2883 57.6049 16.484 60.4853 16.484 60.4853L75.9951 41.5003C75.9951 41.5003 77.3531 37.9514 76.8366 35.6564C76.302 33.2809 75.2382 31.959 73.3883 30.6009C71.8916 29.5022 69.0669 29.0804 69.0669 29.0804C69.0669 29.0804 69.8674 27.8202 70.1012 26.873C70.6832 24.5155 70.3844 22.2398 68.6097 20.6368C66.8453 19.043 62.9764 20.3804 62.9764 20.3804C62.9764 20.3804 62.7586 17.453 61.9014 15.8858C60.4375 13.2095 58.4075 11.8736 55.6902 11.9069C53.1926 11.9376 50.2766 15.3994 50.2766 15.3994C50.2766 15.3994 47.9 12.387 46.0357 10.9492C43.5008 8.99443 41.7662 7.16645 38.8214 7.56662C36.6977 7.85522 35.4359 8.66267 34.1018 10.4773C32.5348 12.6088 32.5394 17.4439 32.5394 17.4439C32.5394 17.4439 30.5698 15.6262 29.0913 15.0019C26.6933 13.9892 24.9235 13.8071 22.6564 14.8778C20.5075 15.8927 19.5393 17.3001 18.5414 19.6633C17.3512 22.4822 18.0896 27.7899 18.0896 27.7899C18.0896 27.7899 15.5775 26.8086 14.1221 27.3133C11.0001 28.396 11.5466 32.7982 12.2061 36.5028C12.6254 38.858 14.638 42.0983 14.638 42.0983C14.638 42.0983 11.067 42.82 9.66309 44.5949C7.6053 47.1963 7.73272 50.2765 8.92028 53.6819Z"
-          fill="#F2E06F"
-        />
-        <path
-          d="M-1.45008 60.3433C-0.306877 62.7832 7.10566 63.7952 7.10566 63.7952L85.3106 42.0573C85.3106 42.0573 87.563 39.4603 87.2537 37.9903C86.9334 36.4687 85.7849 35.7371 83.6464 35.0991C81.9163 34.5829 78.403 34.7227 78.403 34.7227C78.403 34.7227 79.6057 33.7527 80.0441 33.0789C81.135 31.4017 81.0983 29.9121 79.0919 29.0976C77.0971 28.2879 71.9973 29.7706 71.9973 29.7706C71.9973 29.7706 72.1612 27.8293 71.3111 26.9013C69.8593 25.3164 67.4895 24.7204 64.0438 25.1509C60.8767 25.5466 56.6645 28.3188 56.6645 28.3188C56.6645 28.3188 54.1076 26.6443 51.9628 25.9548C49.0466 25.0173 47.1248 24.0451 43.3359 24.7571C40.6034 25.2706 38.8845 26.0045 36.9228 27.4285C34.6185 29.1011 33.8984 32.3608 33.8984 32.3608C33.8984 32.3608 31.6773 31.4309 29.8989 31.2319C27.0146 30.9091 24.801 31.052 21.7696 32.1144C18.8962 33.1215 17.4589 34.2159 15.8405 35.9593C13.9102 38.0388 14.0482 41.507 14.0482 41.507C14.0482 41.507 11.0147 41.2225 9.09604 41.7814C4.98027 42.9803 5.01124 45.8667 5.29006 48.2657C5.46732 49.7909 7.52913 51.6737 7.52913 51.6737C7.52913 51.6737 2.89918 52.6966 0.854963 54.1042C-2.14128 56.1674 -2.44245 58.2253 -1.45008 60.3433Z"
-          fill="#E4C642"
-        />
-      </g>
-      <path
-        d="M10.9502 58.6263C10.3117 55.6751 12.8129 52.577 16.5892 51.6737C20.3657 50.7705 23.9988 52.4014 24.7647 55.3223L10.9502 58.6263Z"
-        fill="#E28913"
-      />
-      <path
-        d="M37.9115 52.1761C37.2711 49.2254 39.673 46.151 43.3264 45.2771C46.9799 44.4033 50.5128 46.0583 51.2768 48.9795L37.9115 52.1761Z"
-        fill="#E28913"
-      />
-      <path
-        d="M24.7677 55.3205C24.1261 52.3699 26.4783 49.3075 30.0705 48.4483C33.6627 47.5892 37.1451 49.2562 37.9079 52.1777L24.7677 55.3205Z"
-        fill="#FFF5E5"
-      />
-      <path
-        d="M77.4511 42.7215C76.6848 39.8009 73.0329 38.1745 69.232 39.0834C65.4309 39.9925 62.9093 43.0955 63.5474 46.0469L77.4511 42.7215Z"
-        fill="#E28913"
-      />
-      <path
-        d="M63.6331 46.0229C62.8735 43.1011 59.566 41.3929 56.1891 42.2003C52.8121 43.008 50.6347 46.0281 51.2794 48.9775L63.6331 46.0229Z"
-        fill="#FFF5E5"
-      />
-      <mask
-        id="mask1_707_15389"
-        style={{ maskType: "alpha" }}
-        maskUnits="userSpaceOnUse"
-        x="10"
-        y="42"
-        width="75"
-        height="86"
-      >
-        <mask
-          id="mask2_707_15389"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x="10"
-          y="49"
-          width="55"
-          height="79"
-        >
-          <path
-            d="M10.9414 58.6016L47.3395 49.8962L64.3204 120.895L37.9206 127.209L10.9414 58.6016Z"
-            fill="#D9D9D9"
-          />
-        </mask>
-        <g mask="url(#mask2_707_15389)">
-          <rect
-            x="5.06295"
-            y="53.8887"
-            width="18.8555"
-            height="85.938"
-            transform="rotate(-17.487 5.06295 53.8887)"
-            fill="#CC322D"
-          />
-          <rect
-            x="23.1664"
-            y="50.1621"
-            width="13.7334"
-            height="85.938"
-            transform="rotate(-17.487 23.1664 50.1621)"
-            fill="#FFF5E5"
-          />
-          <rect
-            x="70.6269"
-            y="126.065"
-            width="13.7334"
-            height="85.938"
-            transform="rotate(165.889 70.6269 126.065)"
-            fill="#CC322D"
-          />
-        </g>
-        <mask
-          id="mask3_707_15389"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x="41"
-          y="42"
-          width="44"
-          height="81"
-        >
-          <path
-            d="M77.4467 42.7002L41.0486 51.4056L58.0295 122.404L84.4294 116.09L77.4467 42.7002Z"
-            fill="#D9D9D9"
-          />
-        </mask>
-        <g mask="url(#mask3_707_15389)">
-          <rect
-            width="18.8555"
-            height="85.938"
-            transform="matrix(-0.98653 0.16358 0.16358 0.98653 80.556 35.8369)"
-            fill="#CC322D"
-          />
-          <rect
-            width="13.7334"
-            height="85.938"
-            transform="matrix(-0.98653 0.16358 0.16358 0.98653 62.7239 40.6992)"
-            fill="#FFF5E5"
-          />
-          <rect
-            width="13.7334"
-            height="85.938"
-            transform="matrix(-0.975179 0.22142 0.22142 0.975179 49.8262 42.7002)"
-            fill="#CC322D"
-          />
-        </g>
-      </mask>
-      <g mask="url(#mask1_707_15389)">
-        <mask
-          id="mask4_707_15389"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x="10"
-          y="49"
-          width="55"
-          height="79"
-        >
-          <path
-            d="M10.9414 58.6016L47.3395 49.8962L64.3204 120.895L37.9206 127.209L10.9414 58.6016Z"
-            fill="#D9D9D9"
-          />
-        </mask>
-        <g mask="url(#mask4_707_15389)">
-          <rect
-            x="5.06295"
-            y="53.8887"
-            width="18.8555"
-            height="85.938"
-            transform="rotate(-17.487 5.06295 53.8887)"
-            fill="#E28913"
-          />
-          <rect
-            x="23.1664"
-            y="50.1621"
-            width="13.7334"
-            height="85.938"
-            transform="rotate(-17.487 23.1664 50.1621)"
-            fill="#FFF5E5"
-          />
-          <rect
-            x="70.6269"
-            y="126.065"
-            width="13.7334"
-            height="85.938"
-            transform="rotate(165.889 70.6269 126.065)"
-            fill="#E28913"
-          />
-        </g>
-        <mask
-          id="mask5_707_15389"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x="41"
-          y="42"
-          width="44"
-          height="81"
-        >
-          <path
-            d="M77.4467 42.7002L41.0486 51.4056L58.0295 122.404L84.4294 116.09L77.4467 42.7002Z"
-            fill="#D9D9D9"
-          />
-        </mask>
-        <g mask="url(#mask5_707_15389)">
-          <rect
-            width="18.8555"
-            height="85.938"
-            transform="matrix(-0.98653 0.16358 0.16358 0.98653 80.556 35.8369)"
-            fill="#E28913"
-          />
-          <rect
-            width="13.7334"
-            height="85.938"
-            transform="matrix(-0.98653 0.16358 0.16358 0.98653 62.7239 40.6992)"
-            fill="#FFF5E5"
-          />
-          <rect
-            width="13.7334"
-            height="85.938"
-            transform="matrix(-0.975179 0.22142 0.22142 0.975179 49.8262 42.7002)"
-            fill="#E28913"
-          />
-        </g>
-        <path
-          d="M16.2655 74.1375C16.2655 74.1375 35.927 101.738 55.0194 105.325C69.4064 108.028 91.9108 98.6821 91.9108 98.6821L95.9276 115.477L25.8277 132.243L16.2655 74.1375Z"
-          fill="#173C62"
-          fillOpacity="0.15"
-          style={{ mixBlendMode: "multiply" }}
-        />
-      </g>
-      <path
-        d="M18.8931 24.138C18.4397 25.1159 19.4227 25.9882 20.3948 26.4539C21.1769 26.8286 22.6152 26.4842 22.6152 26.4842C22.6152 26.4842 23.0779 30.3779 24.424 29.7863C25.5065 29.3105 25.1065 26.8362 25.1065 26.8362C25.1065 26.8362 27.1271 28.0587 27.6798 27.2071C28.1262 26.5195 27.1666 25.1717 27.1666 25.1717C27.1666 25.1717 29.0377 25.1016 29.7015 24.2905C30.2632 23.6042 30.5044 22.9157 30.2134 22.0779C29.8048 20.9016 27.2661 20.8607 27.2661 20.8607C27.2661 20.8607 27.9829 19.9663 27.8193 19.3656C27.6572 18.7703 27.2124 18.4474 26.6018 18.3587C25.5987 18.213 24.9476 20.3588 24.9476 20.3588C24.9476 20.3588 24.0829 19.0309 23.2084 18.9204C22.4256 18.8215 21.8249 19.0466 21.328 19.6595C20.4342 20.7617 22.766 22.9966 22.766 22.9966C22.766 22.9966 21.5169 23.0198 20.7585 23.2401C19.9821 23.4656 19.2332 23.4046 18.8931 24.138Z"
-        fill="#FDF69D"
-      />
-      <path
-        d="M33.7392 19.384C33.0324 19.4619 32.8756 20.3146 32.9446 21.0223C33.0002 21.5917 33.6856 22.286 33.6856 22.286C33.6856 22.286 31.6457 23.8765 32.4392 24.4343C33.0774 24.883 34.3375 23.8125 34.3375 23.8125C34.3375 23.8125 34.3372 25.3704 35.0067 25.3916C35.5472 25.4088 35.9804 24.407 35.9804 24.407C35.9804 24.407 36.6588 25.4393 37.3432 25.5371C37.9224 25.6198 38.3934 25.521 38.7669 25.0707C39.2914 24.4384 38.4478 22.9914 38.4478 22.9914C38.4478 22.9914 39.1974 23.0906 39.4806 22.7932C39.7613 22.4985 39.7917 22.1372 39.6333 21.7622C39.3731 21.1463 37.9396 21.5113 37.9396 21.5113C37.9396 21.5113 38.3939 20.5698 38.1577 20.0385C37.9463 19.5628 37.6142 19.3006 37.0985 19.2294C36.1712 19.1012 35.7058 21.1804 35.7058 21.1804C35.7058 21.1804 35.2662 20.4832 34.8829 20.1304C34.4906 19.7691 34.2694 19.3255 33.7392 19.384Z"
-        fill="#FDF69D"
-      />
-      <path
-        d="M67.1014 31.3794C67.9482 32.0463 67.4663 33.269 66.8101 34.1242C66.2821 34.8121 64.8436 35.1558 64.8436 35.1558C64.8436 35.1558 66.1927 38.8375 64.7247 38.919C63.544 38.9845 62.7811 36.5969 62.7811 36.5969C62.7811 36.5969 61.5325 38.6014 60.6542 38.092C59.945 37.6808 60.1909 36.0447 60.1909 36.0447C60.1909 36.0447 58.4906 36.8288 57.5316 36.4058C56.7202 36.0479 56.1935 35.5431 56.074 34.6643C55.9061 33.4304 58.1516 32.2453 58.1516 32.2453C58.1516 32.2453 57.1077 31.7719 56.9818 31.1622C56.8571 30.558 57.1076 30.0688 57.6119 29.7134C58.4406 29.1297 59.9921 30.7486 59.9921 30.7486C59.9921 30.7486 60.1624 29.1731 60.8923 28.6789C61.5456 28.2365 62.1831 28.1655 62.9036 28.4872C64.1994 29.0658 63.1311 32.1139 63.1311 32.1139C63.1311 32.1139 64.2556 31.5694 65.0316 31.4227C65.826 31.2725 66.4662 30.8792 67.1014 31.3794Z"
-        fill="#F1E06F"
-      />
-      <path
-        d="M52.5213 33.677C52.8663 34.2988 52.2611 34.9196 51.6437 35.2724C51.1469 35.5562 50.1841 35.3982 50.1841 35.3982C50.1841 35.3982 50.0654 37.9821 49.1515 37.6571C48.4165 37.3956 48.5617 35.7486 48.5617 35.7486C48.5617 35.7486 47.2906 36.6493 46.8863 36.1153C46.5598 35.6842 47.1269 34.7516 47.1269 34.7516C47.1269 34.7516 45.8924 34.7947 45.417 34.2927C45.0147 33.8679 44.8232 33.4264 44.9747 32.8613C45.1874 32.0678 46.8559 31.9199 46.8559 31.9199C46.8559 31.9199 46.3416 31.3656 46.4206 30.9625C46.4989 30.5631 46.7761 30.3295 47.1736 30.242C47.8267 30.0983 48.3574 31.4791 48.3574 31.4791C48.3574 31.4791 48.8631 30.5641 49.4332 30.4497C49.9435 30.3473 50.3495 30.4668 50.7057 30.8464C51.3463 31.529 49.9186 33.1107 49.9186 33.1107C49.9186 33.1107 50.7416 33.0663 51.2511 33.1751C51.7727 33.2865 52.2625 33.2106 52.5213 33.677Z"
-        fill="#F1E06F"
-      />
-      <path
-        d="M18.7928 40.7704C19.4026 40.1587 20.3307 40.6565 20.9499 41.2586C21.4482 41.743 21.5873 42.9199 21.5873 42.9199C21.5873 42.9199 24.6434 42.1901 24.5711 43.3661C24.5129 44.3118 22.5416 44.6959 22.5416 44.6959C22.5416 44.6959 24.0202 45.8769 23.5327 46.5283C23.1392 47.0543 21.8601 46.7057 21.8601 46.7057C21.8601 46.7057 22.3253 48.1322 21.8991 48.8558C21.5385 49.4682 21.0875 49.8402 20.377 49.8532C19.3793 49.8715 18.6459 47.9738 18.6459 47.9738C18.6459 47.9738 18.1717 48.7604 17.6747 48.8036C17.1822 48.8465 16.8162 48.6014 16.5806 48.1668C16.1934 47.4528 17.6267 46.3693 17.6267 46.3693C17.6267 46.3693 16.3888 46.0865 16.0637 45.4596C15.7726 44.8983 15.7756 44.3843 16.0989 43.8409C16.6804 42.8637 19.0065 43.9986 19.0065 43.9986C19.0065 43.9986 18.6781 43.0529 18.6339 42.4216C18.5886 41.7754 18.3354 41.2291 18.7928 40.7704Z"
-        fill="#F1E06F"
-      />
-    </svg>
-  );
-};
-export default Popcorn;
