--- apps/main/src/app/lib/Assets/SVG/tent.tsx
@@ -1,379 +1,0 @@
-import React from "react";
-
-const Tent = () => {
-  return (
-    <svg
-      width="1200"
-      height="854"
-      viewBox="0 0 1200 854"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_707_15325)">
-        <rect
-          x="104.143"
-          y="462.001"
-          width="66.1417"
-          height="392.997"
-          fill="#FFF5E5"
-        />
-        <rect
-          x="38"
-          y="462.001"
-          width="66.1417"
-          height="392.997"
-          fill="#FDA829"
-        />
-        <path
-          d="M386.809 203.641C327.454 292.911 221.161 409.427 52.2882 442.84L48.4499 464.43L655.784 470.112C677.731 443.503 689.459 394.48 610.335 339.27C531.201 284.053 464.468 166.829 441.135 120.405C441.923 114.155 435.827 129.919 386.809 203.641Z"
-          fill="#F7DFBC"
-        />
-        <path
-          d="M279.626 324.209C210.516 396.152 75.5906 432.083 0 461.808H66.7323C66.7323 461.808 158.563 454.131 272.835 353.737C324.508 308.339 412.303 173.619 441.437 118.402C424.902 142.713 377.485 222.339 279.626 324.209Z"
-          fill="#E28913"
-        />
-        <path
-          d="M329.823 323.324C267.815 413.678 193.996 444.092 133.464 461.808H200.197C200.197 461.808 289.37 434.052 350.197 330.706C387.401 267.494 421.358 175.981 441.142 119.879C424.606 144.19 411.388 204.471 329.823 323.324Z"
-          fill="#E28913"
-        />
-        <path
-          d="M382.388 319.551C353.011 385.952 327.263 444.151 266.93 461.81H333.662C333.662 461.81 374.891 437.952 404.233 323.03C418.406 267.518 431.398 178.936 441.438 121.062C431.398 149.703 404.233 270.176 382.678 318.896L382.388 319.551Z"
-          fill="#E28913"
-        />
-        <path
-          d="M602.952 324.212C678.663 397.752 792.224 432.086 867.814 461.811H801.082C801.082 461.811 713.976 448.818 599.704 348.425C548.031 303.027 469.192 181.889 442.618 124.311C460.925 147.637 496.554 220.866 602.952 324.212Z"
-          fill="#E28913"
-        />
-        <path
-          d="M551.28 320.37C623.032 414.268 673.819 444.091 734.351 461.807H667.619C667.619 461.807 589.857 429.693 525.296 326.276C493.406 275.193 460.04 184.543 439.666 121.354C456.201 145.666 463.757 205.834 551.28 320.37Z"
-          fill="#E28913"
-        />
-        <path
-          d="M498.426 316.239C534.154 385.174 558.662 426.967 600.886 461.81H534.154C534.154 461.81 502.669 439.134 473.327 324.211C459.154 268.7 449.115 180.412 439.075 122.538C449.115 151.18 473.327 267.814 498.426 316.239Z"
-          fill="#E28913"
-        />
-        <path
-          d="M426.673 289.366C416.338 398.028 408.169 446.945 400.393 461.807L467.126 462.102C467.322 462.004 461.249 430.618 456.496 364.071C451.476 293.795 442.421 178.441 440.846 121.354C440.551 131.492 436.601 184.983 426.673 289.366Z"
-          fill="#E28913"
-        />
-        <path d="M600 868H102L169.929 477L600 477V868Z" fill="#353131" />
-        <g filter="url(#filter0_d_707_15325)">
-          <path
-            d="M104.55 854L38 657.325C47.0196 657.325 64.8869 648.444 96.2616 612.822C141.359 561.62 160.049 512.491 174.756 464L175 479.791L104.55 854Z"
-            fill="#F7DFBC"
-          />
-        </g>
-        <g filter="url(#filter1_d_707_15325)">
-          <path
-            d="M66.7322 461.815H133.464V470.674C133.464 488.612 118.923 503.154 100.984 503.154H99.2125C81.2741 503.154 66.7322 488.612 66.7322 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M133.465 461.815H200.198V470.674C200.198 488.612 185.656 503.154 167.717 503.154H165.946C148.007 503.154 133.465 488.612 133.465 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M200.197 461.815H266.929V470.674C266.929 488.612 252.387 503.154 234.449 503.154H232.677C214.739 503.154 200.197 488.612 200.197 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M266.929 461.815H333.661V470.674C333.661 488.612 319.119 503.154 301.181 503.154H299.409C281.471 503.154 266.929 488.612 266.929 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M333.662 461.815H400.394V470.674C400.394 488.612 385.852 503.154 367.914 503.154H366.142C348.204 503.154 333.662 488.612 333.662 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M400.393 461.815H467.126V470.674C467.126 488.612 452.584 503.154 434.645 503.154H432.874C414.935 503.154 400.393 488.612 400.393 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M467.126 461.815H533.859V470.674C533.859 488.612 519.317 503.154 501.378 503.154H499.607C481.668 503.154 467.126 488.612 467.126 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M533.858 461.815H600.591V470.674C600.591 488.612 586.049 503.154 568.11 503.154H566.339C548.4 503.154 533.858 488.612 533.858 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M0 461.815H66.7323V470.674C66.7323 488.612 52.1904 503.154 34.252 503.154H32.4803C14.5419 503.154 0 488.612 0 470.674V461.815Z"
-            fill="#E28913"
-          />
-        </g>
-      </g>
-      <g clipPath="url(#clip1_707_15325)">
-        <rect
-          width="66.1417"
-          height="392.997"
-          transform="matrix(-1 0 0 1 1095.86 462.001)"
-          fill="#FFF5E5"
-        />
-        <rect
-          width="66.1417"
-          height="392.997"
-          transform="matrix(-1 0 0 1 1162 462.001)"
-          fill="#FDA829"
-        />
-        <path
-          d="M813.191 203.641C872.546 292.911 978.839 409.427 1147.71 442.84L1151.55 464.43L544.216 470.112C522.269 443.503 510.541 394.48 589.665 339.27C668.799 284.053 735.532 166.829 758.865 120.405C758.077 114.155 764.173 129.919 813.191 203.641Z"
-          fill="#F7DFBC"
-        />
-        <path
-          d="M920.374 324.209C989.484 396.152 1124.41 432.083 1200 461.808H1133.27C1133.27 461.808 1041.44 454.131 927.165 353.737C875.492 308.339 787.697 173.619 758.563 118.402C775.098 142.713 822.515 222.339 920.374 324.209Z"
-          fill="#E28913"
-        />
-        <path
-          d="M870.177 323.324C932.185 413.678 1006 444.092 1066.54 461.808H999.803C999.803 461.808 910.63 434.052 849.803 330.706C812.599 267.494 778.642 175.981 758.858 119.879C775.394 144.19 788.612 204.471 870.177 323.324Z"
-          fill="#E28913"
-        />
-        <path
-          d="M817.612 319.551C846.989 385.952 872.737 444.151 933.07 461.81H866.338C866.338 461.81 825.109 437.952 795.767 323.03C781.594 267.518 768.602 178.936 758.562 121.062C768.602 149.703 795.767 270.176 817.322 318.896L817.612 319.551Z"
-          fill="#E28913"
-        />
-        <path
-          d="M597.048 324.212C521.337 397.752 407.776 432.086 332.186 461.811H398.918C398.918 461.811 486.024 448.818 600.296 348.425C651.969 303.027 730.808 181.889 757.382 124.311C739.075 147.637 703.446 220.866 597.048 324.212Z"
-          fill="#E28913"
-        />
-        <path
-          d="M648.72 320.37C576.968 414.268 526.181 444.091 465.649 461.807H532.381C532.381 461.807 610.143 429.693 674.704 326.276C706.594 275.193 739.96 184.543 760.334 121.354C743.799 145.666 736.243 205.834 648.72 320.37Z"
-          fill="#E28913"
-        />
-        <path
-          d="M701.574 316.239C665.846 385.174 641.338 426.967 599.114 461.81H665.846C665.846 461.81 697.331 439.134 726.673 324.211C740.846 268.7 750.885 180.412 760.925 122.538C750.885 151.18 726.673 267.814 701.574 316.239Z"
-          fill="#E28913"
-        />
-        <path
-          d="M773.327 289.366C783.662 398.028 791.831 446.945 799.607 461.807L732.875 462.102C732.678 462.004 738.751 430.618 743.504 364.071C748.524 293.795 757.579 178.441 759.154 121.354C759.449 131.492 763.4 184.983 773.327 289.366Z"
-          fill="#E28913"
-        />
-        <path d="M600 868H1098L1030.07 477L600 477V868Z" fill="#353131" />
-        <g filter="url(#filter2_d_707_15325)">
-          <path
-            d="M1095.45 854L1162 657.325C1152.98 657.325 1135.11 648.444 1103.74 612.822C1058.64 561.62 1039.95 512.491 1025.24 464L1025 479.791L1095.45 854Z"
-            fill="#F7DFBC"
-          />
-        </g>
-        <g filter="url(#filter3_d_707_15325)">
-          <path
-            d="M1133.27 461.815H1066.54V470.674C1066.54 488.612 1081.08 503.154 1099.02 503.154H1100.79C1118.73 503.154 1133.27 488.612 1133.27 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M1066.53 461.815H999.802V470.674C999.802 488.612 1014.34 503.154 1032.28 503.154H1034.05C1051.99 503.154 1066.53 488.612 1066.53 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M999.803 461.815H933.071V470.674C933.071 488.612 947.613 503.154 965.551 503.154H967.323C985.261 503.154 999.803 488.612 999.803 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M933.071 461.815H866.339V470.674C866.339 488.612 880.881 503.154 898.819 503.154H900.591C918.529 503.154 933.071 488.612 933.071 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M866.338 461.815H799.606V470.674C799.606 488.612 814.148 503.154 832.086 503.154H833.858C851.796 503.154 866.338 488.612 866.338 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M799.607 461.815H732.875V470.674C732.875 488.612 747.416 503.154 765.355 503.154H767.127C785.065 503.154 799.607 488.612 799.607 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M732.874 461.815H666.141V470.674C666.141 488.612 680.683 503.154 698.622 503.154H700.393C718.332 503.154 732.874 488.612 732.874 470.674V461.815Z"
-            fill="#F7DFBC"
-          />
-          <path
-            d="M666.142 461.815H599.409V470.674C599.409 488.612 613.951 503.154 631.89 503.154H633.661C651.6 503.154 666.142 488.612 666.142 470.674V461.815Z"
-            fill="#E28913"
-          />
-          <path
-            d="M1200 461.815H1133.27V470.674C1133.27 488.612 1147.81 503.154 1165.75 503.154H1167.52C1185.46 503.154 1200 488.612 1200 470.674V461.815Z"
-            fill="#E28913"
-          />
-        </g>
-      </g>
-      <defs>
-        <filter
-          id="filter0_d_707_15325"
-          x="38"
-          y="464"
-          width="137"
-          height="416.49"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dy="26.4901" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_707_15325"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect1_dropShadow_707_15325"
-            result="shape"
-          />
-        </filter>
-        <filter
-          id="filter1_d_707_15325"
-          x="0"
-          y="461.815"
-          width="600.591"
-          height="54.5839"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dy="13.245" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_707_15325"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect1_dropShadow_707_15325"
-            result="shape"
-          />
-        </filter>
-        <filter
-          id="filter2_d_707_15325"
-          x="1025"
-          y="464"
-          width="137"
-          height="416.49"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dy="26.4901" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_707_15325"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect1_dropShadow_707_15325"
-            result="shape"
-          />
-        </filter>
-        <filter
-          id="filter3_d_707_15325"
-          x="599.409"
-          y="461.815"
-          width="600.591"
-          height="54.5839"
-          filterUnits="userSpaceOnUse"
-          colorInterpolationFilters="sRGB"
-        >
-          <feFlood floodOpacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dy="13.245" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_707_15325"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect1_dropShadow_707_15325"
-            result="shape"
-          />
-        </filter>
-        <clipPath id="clip0_707_15325">
-          <rect width="600" height="854" fill="white" />
-        </clipPath>
-        <clipPath id="clip1_707_15325">
-          <rect
-            width="600"
-            height="854"
-            fill="white"
-            transform="matrix(-1 0 0 1 1200 0)"
-          />
-        </clipPath>
-      </defs>
-      <svg
-        x="300"
-        y="265"
-        width="597"
-        height="93"
-        viewBox="0 0 597 93"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M2.28232 37.9013C1.18703 35.1304 3.22915 32.1273 6.20871 32.1273H85.2844C87.6162 32.1273 89.5064 34.0176 89.5064 36.3493V86.7721C89.5064 89.1038 87.6162 90.9941 85.2844 90.9941H6.37191C3.34895 90.9941 1.30549 87.9101 2.4839 85.1263L11.3592 64.1598C11.7908 63.1401 11.8046 61.9917 11.3976 60.9619L2.28232 37.9013Z"
-          fill="#94181A"
-        />
-        <path
-          d="M84.58 90.648L54.3562 82.2767C49.955 81.0576 50.3538 74.6904 54.8728 74.0302L85.0966 69.6146C87.6443 69.2424 89.929 71.2175 89.929 73.7922V86.5792C89.929 89.37 87.2696 91.393 84.58 90.648Z"
-          fill="#68090B"
-        />
-        <path
-          d="M594.187 37.9482C595.283 35.1773 593.241 32.1742 590.261 32.1742H511.185C508.854 32.1742 506.963 34.0644 506.963 36.3962V86.819C506.963 89.1507 508.854 91.041 511.185 91.041H590.098C593.121 91.041 595.164 87.957 593.986 85.1731L585.111 64.2066C584.679 63.187 584.665 62.0385 585.072 61.0088L594.187 37.9482Z"
-          fill="#94181A"
-        />
-        <path
-          d="M512.312 90.6949L542.535 82.3236C546.937 81.1045 546.538 74.7373 542.019 74.0771L511.795 69.6614C509.247 69.2892 506.963 71.2644 506.963 73.8391V86.6261C506.963 89.4169 509.622 91.4398 512.312 90.6949Z"
-          fill="#68090B"
-        />
-        <path
-          d="M50.9629 13.8682C50.9629 11.643 52.6716 9.80167 54.8925 9.66178C79.1867 8.13148 210.753 0.106227 298.318 0.0882568C386.36 0.0701892 518.656 8.12855 543.031 9.66192C545.252 9.80164 546.963 11.6432 546.963 13.8685V77.6659C546.963 80.0744 545.013 81.9985 542.607 81.8908C517.15 80.7508 385.733 75.0623 298.318 75.0711C211.381 75.0799 80.6955 80.7511 55.3176 81.8903C52.9113 81.9984 50.9629 80.0741 50.9629 77.6654V13.8682Z"
-          fill="#CC322D"
-        />
-      </svg>
-    </svg>
-  );
-};
-
-export default Tent;
