+++ apps/main/src/app/sponsor-us/Sections/SponsorUsStatsComp.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import RibbonTitle from "@repo/ui/RibbonTitle";
 import Image from "next/image";
-import FullCloudBackground from "../../lib/Assets/SVG/SponsorUsAssets/FullCloudBackground";
+import FullCloudBackground from "../../lib/Assets/SVG/SponsorUsPageAssets/FullCloudBackground";
 import useDevice from "@util/hooks/useDevice";
 import clsx from "clsx";
 
@@ -23,7 +23,7 @@ export default function SponsorUsBenefitCardComp() {
     "w-1/2",
     isMobile && "mt-52 mb-10",
     isTablet && "mt-48 mb-8",
-    isDesktop && "mt-64 mb-12",
+    isDesktop && "mt-64 mb-12"
   );
 
   const balloons: BalloonProps[] = [
